C driver FAST restart doesn't work
The c driver program as was present in the tar.gz file on the NREL server runs. It produces a checkpoint after 10 time steps with the names "Test18.T1.10.chkp" and "Test18.T1.10.dll.chkp". I then modified the c file to restart from this file. The program segfaults. Running this in the debugger DDT shows that the c style character array 'CheckpointRootName_C' doesn't get passed properly from FAST_OpFM_Restart to FAST_Restart. While the arrays is correctly shown as 'Test18.T1.10.(JUNKAFTERTHIS)' in FAST_OpFM_Restart, it's completely garbled after it's passed to FAST_Restart. This won't necessarily fix the restart issue if the native version of FAST doesn't restart. However it's a step in the right direction.