This solves the file name issue and the c driver goes through and reads the restart file fully.
However, the code still has a memory issue in setOutputsToFAST routine. I'm not sure it's worth going into more detail here. A more serious problme is that the time step loop in FAST_Prog.c should depend on the values read in from the restart file. This needs to be fixed first. My next job is to make the c driver program give the same error as the Fortran driver program in a restart.