Adding a SuperController module
This module mimicks the OpenFOAM module, but only includes the relevant parts for the inputs and outputs of the supercontroller for each turbine. This module is not similar to other modules in FAST because the supercontroller is fundamentally not designed for one turbine, while all the other module are only for one turbine. This module is designed to only have inputs and outputs along with one parameter 'scOn' that describes whether the module is active or not. I hope that this module works with restart.