Preparing the stage for separating the set of force actuator points from the velocity actuator points
The velocity actuator points will be the Aerodyn nodes while the force actuator points will be the ones at which the actuator body force is applied in a CFD solver. The OpenFOAM module will create a new set of force actuator points by interpolating from the Aerodyn nodes. FAST will compute the forces at the Aerodyn nodes and the OpenFOAM module will interpolate the forces to the new force actuator points.

1. The initialization of the OpenFOAM module now uses the InitInp_AD data as well. This data will be used to initialize the new force actuator points in the same style as the Aerodyn15 init routine. Will support only AD15. This may need to be extended to allow for AD14 use as well.
2. FAST c Interface will separate the existing variables for the number of points in the blade and tower into two: numVelPts and numForcePts. The c interface will also take two extra user defined inputs for each turbine numForcePtsBlade and numForcePtsTwr. My thinking: Even if FAST-Aerodyn has different number of nodes for each blade, it's ok to have the same number of force actuator points for each blade.
3. FAST_Library module will now take two extra inputs in the ExternInitData type: NumActForcePtsBlade and NumActForcePtsTwr and pass them into the OpenFOAM module through the InitInp_OpenFOAM datastructure.