Interpolation of displacements, positions and forces to the new mesh is finally working.
Created functions to compute thrust and torque based on the lumped forces at the actuator points as well. A small test to check the variation of torque and thrust with changing number of actuator force points yields this.

#Torque
#nPoints TorqueX           TorqueY           TorqueZ
10       5197508.477225583 916328.3381480369 -120326.306160452
20       5171440.088057267 929315.7053749202 -122606.0956558565
30       5141865.453733687 937944.8554450116 -139563.8007625903
50       5158888.58451682 922247.8094093908 -85688.31942930883
100      5153602.612688709 922995.6716311999 -110068.5732636223

#Thrust
#nPoints ThrustX           ThrustY           ThrustZ
10       706638.1031341553 -9524.848124980927 -61602.61569786072
20       706638.1093902588 -9524.847578048706 -61602.61482715607
30       706638.1104507446 -9524.848219633102 -61602.61526417732
50       706638.1048469543 -9524.847837209702 -61602.61470675468
100      706638.1077861786 -9524.847942471504 -61602.61520043015

Jason says it could be because of the bug in the line2_to_point_mapping. Need to check this before I finalize this branch.