Interpolated chord to the force nodes and made accessible through a function in the FAST-CPP interface
The chord interpolation uses the BladeProps from the AeroDyn module. I had to copy this data over to the InitOutputType of AeroDyn as the AeroDyn module destroys the InputFileData after initialization.

I also changed the interpolation parameter for the new force node positions to use the same discretization used in the structural model. I had to get a few parameters from the ElastoDyn module. I handled this through additions to the InitOutputType of ElastoDyn. Doesn't work for BeamDyn yet. I have put placeholders to handle BeamDyn in the future.