Moving supercontroller data into a separate class stored as an object within the FAST - C interface
The previous FAST - C  interface stores the states of the supercontroller within a separate supercontroller class with no control over the data. This does not conform to the FASTv8 interface because the supercontroller class cannot correct the supercontroller state once updateStates is called. Hence the FAST c Interface class needs to store all the super controller states along with a duplicate copy for the prediction at the next time and pass it to the supercontroller functions everytime. Finally, when we are sure that enough corrections have been made, the states have to be updated to the next time step. The steps to effect this transition are:

1. Move all existing supercontroller states storage into FAST cInterface, create extra storage to store the predicted states at the next time step.
2. Change updateStates and calcOutputs to pass the states into the function.